% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastphase.R
\name{assemble_pEmit}
\alias{assemble_pEmit}
\title{Compute the genotype emission distributions based on the fastPhase HMM}
\usage{
assemble_pEmit(theta)
}
\description{
Compute the genotype emission distributions based on the fastPhase HMM
}
\keyword{internal}
