% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastphase.R
\name{assemble_pEmit_phased}
\alias{assemble_pEmit_phased}
\title{Compute the haplotype emission distributions based on the fastPHASE HMM}
\usage{
assemble_pEmit_phased(theta)
}
\description{
Compute the haplotype emission distributions based on the fastPHASE HMM
}
\keyword{internal}
