% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastphase.R
\name{assemble_pInit}
\alias{assemble_pInit}
\title{Compute the genotype initial distributions based on the fastPhase HMM}
\usage{
assemble_pInit(alpha)
}
\description{
Compute the genotype initial distributions based on the fastPhase HMM
}
\keyword{internal}
