% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastphase.R
\name{assemble_pInit_phased}
\alias{assemble_pInit_phased}
\title{Compute the haplotype initial distributions based on the fastPHASE HMM}
\usage{
assemble_pInit_phased(alpha)
}
\description{
Compute the haplotype initial distributions based on the fastPHASE HMM
}
\keyword{internal}
