% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastphase.R
\name{assemble_Q}
\alias{assemble_Q}
\title{Compute the genotype transition matrices based on the fastPHASE HMM}
\usage{
assemble_Q(Q1)
}
\description{
Compute the genotype transition matrices based on the fastPHASE HMM
}
\keyword{internal}
