% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{fwHaplotypes}
\alias{fwHaplotypes}
\title{Forward-backward sampling for hidden Markov models for haplotypes}
\usage{
fwHaplotypes(X, r, alpha, theta, seed = 123, cluster = NULL,
  display_progress = FALSE)
}
\arguments{
\item{X}{a binary matrix of size n-by-p containing the original variables.}

\item{r}{a vector of length p containing the "r" parameters estimated by fastPHASE.}

\item{alpha}{a matrix of size p-by-K containing the "alpha" parameters estimated by fastPHASE.}

\item{theta}{a matrix of size p-by-K containing the "theta" parameters estimated by fastPHASE.}

\item{seed}{an integer random seed (default: 123).}

\item{cluster}{a computing cluster object created by \link[parallel]{makeCluster} (default: NULL).}

\item{display_progress}{whether to show progress bar (default: FALSE).}
}
\value{
A binary matrix of size n-by-p containing the knockoff variables.
}
\description{
This function samples the latent Markov chain of a hidden Markov model given the observed values,
given an HMM for phased haplotypes.
}
\examples{
# Problem size
p = 10
n = 100
# Load HMM to generate data
r_file = system.file("extdata", "haplotypes_rhat.txt", package = "SNPknock")
alpha_file = system.file("extdata", "haplotypes_alphahat.txt", package = "SNPknock")
theta_file = system.file("extdata", "haplotypes_thetahat.txt", package = "SNPknock")
char_file = system.file("extdata", "haplotypes_origchars", package = "SNPknock")
hmm.data = loadHMM(r_file, alpha_file, theta_file, char_file, compact=FALSE, phased=TRUE)
hmm.data$Q = hmm.data$Q[1:(p-1),,]
hmm.data$pEmit = hmm.data$pEmit[1:p,,]
# Sample X from this HMM
X = sampleHMM(hmm.data$pInit, hmm.data$Q, hmm.data$pEmit, n=n)
# Load HMM to generate knockoffs
hmm = loadHMM(r_file, alpha_file, theta_file, char_file)
hmm$r = hmm$r[1:p]
hmm$alpha = hmm$alpha[1:p,]
hmm$theta = hmm$theta[1:p,]
# Forward-backward sampling
H = fwHaplotypes(X, hmm$r, hmm$alpha, hmm$theta)

}
\references{
\insertRef{sesia2019multi}{SNPknock}
}
\seealso{
Other hmm: \code{\link{fwGenotypes}}, \code{\link{fwHMM}}
}
\concept{hmm}
