% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fwHaplotypes_wrapper}
\alias{fwHaplotypes_wrapper}
\title{Wrapper for HMM forward-backward sampling (optimized for haplotypes)}
\usage{
fwHaplotypes_wrapper(X_, r_, alpha_, theta_, groups_, n_, p_, seed_,
  display_progress_)
}
\description{
Wrapper for HMM forward-backward sampling (optimized for haplotypes)
}
\keyword{internal}
