% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fwHMM_wrapper}
\alias{fwHMM_wrapper}
\title{Wrapper for HMM forward-backward sampling}
\usage{
fwHMM_wrapper(X_, pInit_, Q_, pEmit_, n_, p_, K_, M_, seed_, G_,
  display_progress_)
}
\description{
Wrapper for HMM forward-backward sampling
}
\keyword{internal}
