% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenotypeModel_wrapper}
\alias{GenotypeModel_wrapper}
\title{Wrapper for Genotype model}
\usage{
GenotypeModel_wrapper(X_, r_, alpha_, theta_, n_, p_, seed_, display_progress_)
}
\description{
Wrapper for Genotype model
}
\keyword{internal}
