% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GroupGenotypes_wrapper}
\alias{GroupGenotypes_wrapper}
\title{Wrapper for GroupGenotypes}
\usage{
GroupGenotypes_wrapper(X_, r_, alpha_, theta_, groups_, n_, p_, seed_,
  display_progress_)
}
\description{
Wrapper for GroupGenotypes
}
\keyword{internal}
