% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GroupHaplotypes_wrapper}
\alias{GroupHaplotypes_wrapper}
\title{Wrapper for GroupHaplotypes}
\usage{
GroupHaplotypes_wrapper(X_, r_, alpha_, theta_, groups_, n_, p_, seed_,
  display_progress_)
}
\description{
Wrapper for GroupHaplotypes
}
\keyword{internal}
