% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HaplotypeModel_wrapper}
\alias{HaplotypeModel_wrapper}
\title{Wrapper for Haplotype model}
\usage{
HaplotypeModel_wrapper(X_, r_, alpha_, theta_, n_, p_, seed_, display_progress_)
}
\description{
Wrapper for Haplotype model
}
\keyword{internal}
