% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knockoffDMC_wrapper}
\alias{knockoffDMC_wrapper}
\title{Wrapper for DMC knockoffs}
\usage{
knockoffDMC_wrapper(X_, pInit_, Q_, n_, p_, K_, seed_)
}
\description{
Wrapper for DMC knockoffs
}
\keyword{internal}
