% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knockoffHMM_wrapper}
\alias{knockoffHMM_wrapper}
\title{Wrapper for HMM knockoffs}
\usage{
knockoffHMM_wrapper(X_, pInit_, Q_, pEmit_, n_, p_, K_, M_, seed_)
}
\description{
Wrapper for HMM knockoffs
}
\keyword{internal}
